# docs/swagger.yaml
definitions:
  consts.ConfigStatus:
    enum:
    - 0
    type: integer
    x-enum-comments:
      ConfigStatusEnable: 启用
    x-enum-varnames:
    - ConfigStatusEnable
...
  consts.DictStatus:
    enum:
    - 0
    type: integer
    x-enum-comments:
      DictStatusEnable: 启用
    x-enum-varnames:
    - DictStatusEnable
  consts.DictType:
    enum:
    - 0
    type: integer
    x-enum-comments:
      DictTypeInt: int
    x-enum-varnames:
    - DictTypeInt
...
  consts.UserType:
    enum:
    - 0
    type: integer
    x-enum-comments:
      UserTypeGeneral: 普通用户
    x-enum-varnames:
    - UserTypeGeneral
  dto.AuthMenuVo:
    properties:
      children:
        description: 子菜单
        items:
          $ref: '#/definitions/dto.AuthMenuVo'
        type: array
      component:
        description: 组件路径
        type: string
      meta:
        allOf:
        - $ref: '#/definitions/dto.MenuRouterMetaVo'
        description: Mate
      name:
        description: 路由名称（TODO：暂时不清楚前端这个字段有什么用处）
        type: string
      path:
        description: 路由地址
        type: string
    type: object
  dto.CaptchaVo:
    properties:
      captchaBase64:
        type: string
      captchaId:
        type: string
    type: object
  dto.DictItemVo:
    properties:
      dictItemName:
        description: 选项名称
        type: string
      dictItemValue:
        description: 选项值
    type: object
  dto.LoginDto:
    properties:
      captchaCode:
        description: 验证码
        type: string
      captchaId:
        description: 验证码ID
        type: string
      password:
        description: 密码
        type: string
      userName:
        description: 用户账号
        type: string
    type: object
  dto.MenuRouterMetaVo:
    properties:
      icon:
        description: 菜单图标
        type: string
      isAffix:
        description: 是否固定在 tagsView 栏上
        type: boolean
      isHide:
        description: 是否隐藏
        type: boolean
      isIframe:
        description: 是否内嵌
        type: boolean
      isKeepAlive:
        description: 是否缓存
        type: boolean
      isLink:
        description: 外部地址
        type: string
      title:
        description: 菜单标题
        type: string
    type: object
...
  dto.SysConfigAddDto:
    properties:
      configKey:
        description: 配置键名
        type: string
      configName:
        description: 配置名称
        type: string
      configValue:
        description: 配置键值
        type: string
      remark:
        description: 备注
        type: string
      status:
        allOf:
        - $ref: '#/definitions/consts.ConfigStatus'
        description: 配置状态（0正常 1停用）
    type: object
  dto.SysConfigUpdateDto:
    properties:
      configKey:
        description: 配置键名
        type: string
      configName:
        description: 配置名称
        type: string
      configValue:
        description: 配置键值
        type: string
      id:
        description: 编号
        type: integer
      remark:
        description: 备注
        type: string
      status:
        allOf:
        - $ref: '#/definitions/consts.ConfigStatus'
        description: 配置状态（0正常 1停用）
    type: object
  dto.SysConfigVo:
    properties:
      configKey:
        description: 配置键名
        type: string
      configName:
        description: 配置名称
        type: string
      configValue:
        description: 配置键值
        type: string
      createdAt:
        description: 创建时间
        type: string
      id:
        description: 配置id
        type: integer
      remark:
        description: 备注
        type: string
      status:
        allOf:
        - $ref: '#/definitions/consts.ConfigStatus'
        description: 配置状态（0正常 1停用）
      updatedAt:
        description: 更新时间
        type: string
    type: object
  dto.SysDictAddDto:
    properties:
      dictCode:
        description: 字典代码
        type: string
      dictName:
        description: 字典名称
        type: string
      dictType:
        allOf:
        - $ref: '#/definitions/consts.DictType'
        description: 字典类型（0int，1string）
      remark:
        description: 备注
        type: string
      status:
        allOf:
        - $ref: '#/definitions/consts.DictStatus'
        description: 字典状态（0正常 1停用）
    type: object
...
  dto.SysDictUpdateDto:
    properties:
      dictCode:
        description: 字典代码
        type: string
      dictName:
        description: 字典名称
        type: string
      dictType:
        allOf:
        - $ref: '#/definitions/consts.DictType'
        description: 字典类型（0int，1string）
      id:
        description: 编号
        type: integer
      remark:
        description: 备注
        type: string
      status:
        allOf:
        - $ref: '#/definitions/consts.DictStatus'
        description: 字典状态（0正常 1停用）
    type: object
  dto.SysDictVo:
    properties:
      createdAt:
        description: 创建时间
        type: string
      dictCode:
        description: 字典代码
        type: string
      dictName:
        description: 字典名称
        type: string
      dictType:
        allOf:
        - $ref: '#/definitions/consts.DictType'
        description: 字典类型（0int，1string）
      id:
        description: 编号
        type: integer
      items:
        description: 选项
        items:
          $ref: '#/definitions/dto.DictItemVo'
        type: array
      remark:
        description: 备注
        type: string
      status:
        allOf:
        - $ref: '#/definitions/consts.DictStatus'
        description: 字典状态（0正常 1停用）
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
  dto.UpdatePwdDto:
    properties:
      newPassword:
        description: 新密码
        type: string
      password:
        description: 密码
        type: string
    type: object
  dto.UserInfoVo:
    properties:
      avatar:
        description: 头像地址
        type: string
      email:
        description: 用户邮箱
        type: string
      gender:
        description: 用户性别（0未知，1男，2女）
        type: integer
      id:
        description: 编号
        type: integer
      nickName:
        description: 用户昵称
        type: string
      permission:
        description: 权限
        items:
          type: string
        type: array
      phone:
        description: 手机号码
        type: string
      remark:
        description: 备注
        type: string
      status:
        description: 帐号状态（0正常 1停用）
        type: integer
      userName:
        description: 用户账号
        type: string
      userType:
        allOf:
        - $ref: '#/definitions/consts.UserType'
        description: 用户类型（0普通账号，1超级管理员）
    type: object
  request.IdInfoDto:
    properties:
      id:
        description: id编号
        type: integer
    type: object
  response.JsonResult:
    properties:
      code:
        description: 状态码（架构层面：中间件/异常处理等）
        type: integer
      data:
        description: 数据
      message:
        description: 消息
        type: string
      success:
        description: 是否成功（业务层面）
        type: boolean
    type: object
  response.PageResult:
    properties:
      list:
        description: 数据
      totalCount:
        description: 总条数
        type: integer
    type: object
info:
  contact: {}
  description: go-fast-admin 接口文档
  title: 系统模块
  version: 1.0.0
paths:
  /system/auth/captcha:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  $ref: '#/definitions/dto.CaptchaVo'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 生成验证码
      tags:
      - 授权
  /system/auth/login:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.LoginDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 用户登录
      tags:
      - 授权
  /system/auth/menu:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.AuthMenuVo'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取用户菜单（树状）
      tags:
      - 授权
  /system/auth/updatePwd:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePwdDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 更新密码
      tags:
      - 授权
  /system/auth/userInfo:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserInfoVo'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取用户信息
      tags:
      - 授权
  /system/config/add:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SysConfigAddDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 添加配置
      tags:
      - 配置
  /system/config/delete:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdInfoDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 删除配置
      tags:
      - 配置
  /system/config/detail:
    get:
      parameters:
      - description: id编号
        in: query
        name: id
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  $ref: '#/definitions/dto.SysConfigVo'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取配置详情
      tags:
      - 配置
...
  /system/config/query:
    get:
      parameters:
      - description: 配置名称
        in: query
        name: configName
        type: string
      - description: 页码
        in: query
        name: pageIndex
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/dto.SysConfigVo'
                        type: array
                    type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: 配置查询
      tags:
      - 配置
  /system/config/update:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SysConfigUpdateDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 更新配置
      tags:
      - 配置
  /system/dict/add:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SysDictAddDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 添加字典
      tags:
      - 字典
  /system/dict/delete:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdInfoDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 删除字典
      tags:
      - 字典
  /system/dict/detail:
    get:
      parameters:
      - description: id编号
        in: query
        name: id
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  $ref: '#/definitions/dto.SysDictVo'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取字典详情
      tags:
      - 字典
  /system/dict/list:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.SysDictVo'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: 字典列表
      tags:
      - 字典
  /system/dict/query:
    get:
      parameters:
      - description: 字典名称
        in: query
        name: dictName
        type: string
      - description: 页码
        in: query
        name: pageIndex
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/dto.SysDictVo'
                        type: array
                    type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: 字典查询
      tags:
      - 字典
  /system/dict/update:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SysDictUpdateDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 更新字典
      tags:
      - 字典
...
  /system/monitor/cacheKeys:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取缓存key
      tags:
      - 监控
  /system/monitor/cacheValue:
    get:
      parameters:
      - description: 缓存Key
        in: query
        name: key
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取缓存value
      tags:
      - 监控
  /system/monitor/deleteCache:
    post:
      parameters:
      - description: 缓存Key
        in: query
        name: key
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 删除缓存
      tags:
      - 监控
  /system/monitor/serverInfo:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取服务器信息
      tags:
      - 监控
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"